# Avoid double buffering when writing CSV/TBL