`generate_in_chunks` now already buffers data in large mutli-MB memory buffers

Using `BufWriter` when writing such chunks simply results in copying the data again. Let's not use the buf write when possible

I noticed this while doing profiling on https://github.com/clflushopt/tpchgen-rs/issues/76

There were 32 buffers allocated instead of just 16 (for each core). 

After this change I verified that there are only 16 x 16mB buffers (so it saves about 100MB of peak memory use)

I also measured performance by generating LineItem SF 100 (80GB of data) and I saw a slight increase (the numbers fluctuated a bit so I am not sure how real)

| branch | Speed | Time |
|--------|--------|
| Main |  (3.04 GB/sec) | 24.5 | 
| This PR | (3.22 GB/sec) | 23.2 | 

Main
```shell
andrewlamb@Andrews-MacBook-Pro-2:~/Software/tpchgen-rs$ RUST_LOG=debug cargo run --release --  --format=csv --tables=line-item -s 100 --output-dir=/tmp/tpchgen-rs
   Compiling tpchgen-cli v0.1.0 (/Users/andrewlamb/Software/tpchgen-rs/tpchgen-cli)
    Finished `release` profile [optimized + debuginfo] target(s) in 5.40s
     Running `target/release/tpchgen-cli --format=csv --tables=line-item -s 100 --output-dir=/tmp/tpchgen-rs`
[2025-03-29T10:55:04Z DEBUG tpchgen_cli] Logging configured from environment variables
[2025-03-29T10:55:04Z DEBUG tpchgen_cli] Creating distributions and text pool
[2025-03-29T10:55:05Z INFO  tpchgen_cli] Created static distributions and text pools in 854.446292ms
[2025-03-29T10:55:05Z INFO  tpchgen_cli] Writing table lineitem (SF=100) to lineitem.csv
[2025-03-29T10:55:05Z DEBUG tpchgen_cli] Generating 4883 parts in total
[2025-03-29T10:55:06Z DEBUG tpchgen_cli::generate] Using 16 threads
[2025-03-29T10:55:31Z DEBUG tpchgen_cli::generate] waiting for writer task to complete
[2025-03-29T10:55:31Z INFO  tpchgen_cli::statistics] Created 74.67 GB in 24.558035583s (3.04 GB/sec)
[2025-03-29T10:55:31Z DEBUG tpchgen_cli::statistics] Wrote 80180650462 bytes in 4883 buffers  15.66 MB/buffers
[2025-03-29T10:55:31Z INFO  tpchgen_cli] Generation complete!
```


This PR
```shell
andrewlamb@Andrews-MacBook-Pro-2:~/Software/tpchgen-rs$ RUST_LOG=debug cargo run --release --  --format=csv --tables=line-item -s 100 --output-dir=/tmp/tpchgen-rs
    Finished `release` profile [optimized + debuginfo] target(s) in 0.10s
     Running `target/release/tpchgen-cli --format=csv --tables=line-item -s 100 --output-dir=/tmp/tpchgen-rs`
[2025-03-29T10:52:13Z DEBUG tpchgen_cli] Logging configured from environment variables
[2025-03-29T10:52:13Z DEBUG tpchgen_cli] Creating distributions and text pool
[2025-03-29T10:52:14Z INFO  tpchgen_cli] Created static distributions and text pools in 820.529458ms
[2025-03-29T10:52:14Z INFO  tpchgen_cli] Writing table lineitem (SF=100) to lineitem.csv
[2025-03-29T10:52:14Z DEBUG tpchgen_cli] Generating 4883 parts in total
[2025-03-29T10:52:14Z DEBUG tpchgen_cli::generate] Using 16 threads
[2025-03-29T10:52:37Z DEBUG tpchgen_cli::generate] waiting for writer task to complete
[2025-03-29T10:52:37Z INFO  tpchgen_cli::statistics] Created 74.67 GB in 23.181299s (3.22 GB/sec)
[2025-03-29T10:52:37Z DEBUG tpchgen_cli::statistics] Wrote 80180650462 bytes in 4883 buffers  15.66 MB/buffers
[2025-03-29T10:52:37Z INFO  tpchgen_cli] Generation complete!
```